+++ sdk/azcore/arm/internal/resource/resource_identifier_test.go
@@ -4,7 +4,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-package arm
+package resource
 
 import "testing"
